fix(auth): gate /account and /dashboard against anon/non-admin access
app/account/page.tsx rendered an empty server shell to anonymous visitors
(the data-fetch try/catch swallowed everything and AccountClient still
rendered) instead of bouncing to login like its child pages. Resolve the
user before the try and redirect to /auth/login?next=/account when absent.

app/dashboard/page.tsx was a client component that rendered the analytics
skeleton to anyone — proxy.ts noindexes /dashboard but does not auth it,
and /api/analytics-dashboard is admin-only, so non-admins saw a skeleton
then an error state. Split the client UI into DashboardClient.tsx and add
a server page that redirects anon and signed-in non-admins to login,
mirroring the getAdminEmails() allowlist (case-insensitive).

Focused tests cover the anon /account redirect and the /dashboard
anon/non-admin/admin matrix.
